requirements.txt
#google-genai
gTTS
#moviepy
Pillow
#requests





#torch
#diffusers
#transformers
#TTS
#PyYAML
#numpy
#opencv-python
